Fixing gyro and staring implementing menu

+++ empty_CC2650STK_TI_2023/util/avgArray.c
@@ -7,15 +7,16 @@
 
 #include <stdlib.h>
 #include "avgArray.h"
+#include "Board.h"
 
-struct avgArray* createAvgArray(int size, float startVal) {
+struct avgArray* createAvgArray(uint8_t size, float startVal) {
     // Create a pointer to the array, and allocate memory for it.
     struct avgArray* arr = (struct avgArray*)malloc(sizeof(struct avgArray));
 
     // Also allocate memory for the array inside the struct.
     arr->arr = (float*)malloc(sizeof(float) * size);
 
-    int i;
+    uint8_t i;
     // Initialize the array with given values.
     for (i = 0; i < size; i++) {
         *(arr->arr + i) = startVal;
@@ -45,9 +46,9 @@ void updateAvgArray(struct avgArray *arr, float val) {
     arr->avg = calcAvg(arr->arr, arr->size);
 }
 
-float calcAvg(float* arr, int size) {
+float calcAvg(float* arr, uint8_t size) {
     float avg = 0.0;
-    int i;
+    uint8_t i;
 
     // Loop though the array and add to the average.
     for (i = 0; i < size; i++) {
